+++ mvo/src/SlidingWindow.hpp
@@ -23,9 +23,9 @@ public:
   SlidingWindow(int len);
   ~SlidingWindow();
 
-  void newWindow(std::vector<cv::Point2f> trackedFeaturesNow, std::vector<cv::Point2f>& trackedFeaturesBefore,
+  void newWindow(std::vector<cv::Point2d> trackedFeaturesNow, std::vector<cv::Point2d>& trackedFeaturesBefore,
                  std::vector<unsigned char>& found, cv::Mat image);
-  void addFeaturesToCurrentWindow(std::vector<cv::Point2f> features);
+  void addFeaturesToCurrentWindow(std::vector<cv::Point2d> features);
 
   void addTransformationToCurrentWindow(cv::Mat translation, cv::Mat rotation);
 
@@ -33,7 +33,7 @@ public:
    * past = 0, means Current Windows Features
    * When no Window at this time, nullptr is returned
    */
-  std::vector<cv::Point2f>* getFeatures(int past);
+  std::vector<cv::Point2d>* getFeatures(int past);
   cv::Mat* getImage(int past);
   cv::Mat getPosition(int past);
 
@@ -42,8 +42,8 @@ public:
  * Window2 is timely after Window1. While 0 means NOW, and the Maximum is _length.
  * Usually Window 2 is 0.
  */
-  void getCorrespondingFeatures(int window1Index, int window2Index, std::vector<cv::Point2f>& features1,
-                                std::vector<cv::Point2f>& features2);
+  void getCorrespondingFeatures(int window1Index, int window2Index, std::vector<cv::Point2d>& features1,
+                                std::vector<cv::Point2d>& features2);
 
   void getCorrespondingPosition(int window1Index, int window2Index, cv::Mat & position1, cv::Mat & position2, cv::Mat &rotation1, cv::Mat &rotation2);
